Fixed compilation errors on 64-bit platform
std::min() expects the type of both parameters
to be the same. On a 32-bit platform size_t and
unsigned int are the same, but on 64-bit they
are actually different types, so a cast is needed.

Open-webOS-DCO-1.0-Signed-off-by: Teppo Pennanen <teppo.pennanen@palm.com>

+++ imap/src/sync/UIDMap.cpp
@@ -24,7 +24,7 @@ using namespace std;
 UIDMap::UIDMap(const vector<UID>& initialUidList, unsigned int msgCount, unsigned int maxSize)
 {
 	// Copy the uids (must already be sorted)
-	unsigned int listSize = maxSize > 0 ? std::min(maxSize, initialUidList.size()) : initialUidList.size();
+	unsigned int listSize = maxSize > 0 ? std::min(maxSize, (unsigned int)(initialUidList.size())) : initialUidList.size();
 
 	m_uids.insert(m_uids.begin(), initialUidList.end() - listSize, initialUidList.end());
